chore : change structure of the project
