+++ src/adpaters/repositories/prisma/account/db-get-account-by-email.ts
@@ -1,6 +1,6 @@
 import { Account } from '@prisma/client';
 import { FindAccountByEmailRepository } from '../../db/account/get-account-by-email-repository';
-import { prisma } from '../../../database/prisma';
+import { prisma } from '../../../../infra/lib/prisma';
 
 export class DbGetAccountByEmail implements FindAccountByEmailRepository {
   async findByEmail(email: string): Promise<Account | null> {
