+++ src/adpaters/repositories/prisma/cards/db-get-all-cards.ts
@@ -1,6 +1,6 @@
 import { Card } from '@prisma/client';
 import { AllCardsRepository } from '../../db/card/all-cards-repository';
-import { prisma } from '../../../database/prisma';
+import { prisma } from '../../../../infra/lib/prisma';
 
 export class DbGetAllCards implements AllCardsRepository {
   async all(id_account: string): Promise<Card[] | null> {
