+++ src/adpaters/repositories/prisma/transactions/db-all-transactions.ts
@@ -1,6 +1,6 @@
 import { Transactions } from '@prisma/client';
 import { GetAllTransactionsRepository } from '../../db/transactions/all-transactions';
-import { prisma } from '../../../database/prisma';
+import { prisma } from '../../../../infra/lib/prisma';
 
 export class DbGetAllTransactions implements GetAllTransactionsRepository {
   async all(account_id: string): Promise<Transactions[] | null> {
