+++ src/adpaters/repositories/prisma/transactions/db-create-transaction.ts
@@ -1,6 +1,6 @@
 import { Prisma, Transactions } from '@prisma/client';
 import { CreateTransactionRepository } from '../../db/transactions/create-transaction-repository';
-import { prisma } from '../../../database/prisma';
+import { prisma } from '../../../../infra/lib/prisma';
 
 export class DbCreateTransaction implements CreateTransactionRepository {
   async add(
