+++ src/infra/helpers/validations/validate-password.ts
