+++ src/infra/http/controllers/cards/all.ts
@@ -2,23 +2,23 @@ import { Request, Response } from 'express';
 import { HandleRequestController } from '../account/add';
 import { z } from 'zod';
 import { HttpStatus } from '../../../helpers/http/status-code';
-import { makeCardUseCase } from '../../../use-cases/factories/card';
+import { makeCardUseCase } from '../../../../use-cases/factories/card';
 
-export class GetCard implements HandleRequestController {
+export class GetAllCards implements HandleRequestController {
   async handle(req: Request, res: Response): Promise<Response> {
-    const listCardByIdUseCase = await makeCardUseCase();
-
+    const makeListAllCards = await makeCardUseCase();
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
+
     const { id } = paramsSchema.parse(req.params);
 
     try {
-      const card = await listCardByIdUseCase.listCard(id);
+      const cards = await makeListAllCards.listAllCards(id);
 
       return res.status(HttpStatus.Ok).json({
-        msg: 'card found successfully',
-        card,
+        msg: 'All cards user ',
+        cards,
       });
     } catch (error) {
       return res.status(HttpStatus.InternalSeverError).json({
@@ -28,4 +28,4 @@ export class GetCard implements HandleRequestController {
   }
 }
 
-export const getCardByIdHandler = new GetCard().handle;
+export const getAllCardByUserIdHandler = new GetAllCards().handle;
